+++ js/sdk/package.json
@@ -16,12 +16,14 @@
     "assert": "^2.0.0",
     "axios": "^0.26.0",
     "borsh": "^0.4.0",
+    "browser-id3-writer": "^4.4.0",
     "coingecko-api": "^1.0.10",
     "crypto-js": "^4.1.1",
     "decimal.js": "^10.4.3",
     "encoding": "^0.1.13",
     "eslint": "^7.27.0",
     "events": "^3.3.0",
+    "file-saver": "^2.0.5",
     "formik": "^2.2.9",
     "https": "^1.0.0",
     "idb-keyval": "^5.0.6",
